Separate automaxprocs and maxprocs packages
