+++ maxprocs/example_test.go
@@ -18,6 +18,29 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// Package automaxprocs lets Go programs easily configure runtime.GOMAXPROCS
-// to match the configured Linux CPU quota.
-package automaxprocs // import "go.uber.org/automaxprocs"
+package maxprocs_test
+
+import (
+	"log"
+
+	"go.uber.org/automaxprocs/maxprocs"
+)
+
+func Example() {
+	undo, err := maxprocs.Set()
+	defer undo()
+	if err != nil {
+		log.Fatalf("failed to set GOMAXPROCS: %v", err)
+	}
+	// Insert your application logic here.
+}
+
+func ExampleLogger() {
+	// By default, Set doesn't output any logs. You can enable logging by
+	// supplying a printf implementation.
+	undo, err := maxprocs.Set(maxprocs.Logger(log.Printf))
+	defer undo()
+	if err != nil {
+		log.Fatalf("failed to set GOMAXPROCS: %v", err)
+	}
+}
